Split out an HTTP client so the API class can be only API methods
