+++ examples/plain-ruby.rb
@@ -2,10 +2,11 @@ require_relative 'helpers'
 
 # An executable specification that demonstrates how to use the Nylas Ruby SDK to interact with the API. It
 # follows the rough structure of the [Nylas API Reference](https://docs.nylas.com/reference).
-
 api = Nylas::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'], access_token:
                      ENV['NYLAS_ACCESS_TOKEN'])
 
+# Retrieving the account information for given access token
+demonstrate { api.current_account.to_h }
 # Retrieving a count of contacts
 demonstrate { api.contacts.count }
